Refactor DirectedDiscussion so it can be used instructor mode
